+++ src/Service/Athena/src/Input/GetDatabaseInput.php
@@ -28,9 +28,18 @@ final class GetDatabaseInput extends Input
     private $databaseName;
 
     /**
+     * The name of the workgroup for which the metadata is being fetched. Required if requesting an IAM Identity Center
+     * enabled Glue Data Catalog.
+     *
+     * @var string|null
+     */
+    private $workGroup;
+
+    /**
      * @param array{
      *   CatalogName?: string,
      *   DatabaseName?: string,
+     *   WorkGroup?: null|string,
      *   '@region'?: string|null,
      * } $input
      */
@@ -38,6 +47,7 @@ final class GetDatabaseInput extends Input
     {
         $this->catalogName = $input['CatalogName'] ?? null;
         $this->databaseName = $input['DatabaseName'] ?? null;
+        $this->workGroup = $input['WorkGroup'] ?? null;
         parent::__construct($input);
     }
 
@@ -45,6 +55,7 @@ final class GetDatabaseInput extends Input
      * @param array{
      *   CatalogName?: string,
      *   DatabaseName?: string,
+     *   WorkGroup?: null|string,
      *   '@region'?: string|null,
      * }|GetDatabaseInput $input
      */
@@ -61,6 +72,11 @@ final class GetDatabaseInput extends Input
     public function getDatabaseName(): ?string
     {
         return $this->databaseName;
+    }
+
+    public function getWorkGroup(): ?string
+    {
+        return $this->workGroup;
     }
 
     /**
@@ -102,6 +118,13 @@ final class GetDatabaseInput extends Input
         return $this;
     }
 
+    public function setWorkGroup(?string $value): self
+    {
+        $this->workGroup = $value;
+
+        return $this;
+    }
+
     private function requestBody(): array
     {
         $payload = [];
@@ -113,6 +136,9 @@ final class GetDatabaseInput extends Input
             throw new InvalidArgument(sprintf('Missing parameter "DatabaseName" for "%s". The value cannot be null.', __CLASS__));
         }
         $payload['DatabaseName'] = $v;
+        if (null !== $v = $this->workGroup) {
+            $payload['WorkGroup'] = $v;
+        }
 
         return $payload;
     }