+++ src/Service/Athena/src/Input/ListDatabasesInput.php
@@ -35,10 +35,19 @@ final class ListDatabasesInput extends Input
     private $maxResults;
 
     /**
+     * The name of the workgroup for which the metadata is being fetched. Required if requesting an IAM Identity Center
+     * enabled Glue Data Catalog.
+     *
+     * @var string|null
+     */
+    private $workGroup;
+
+    /**
      * @param array{
      *   CatalogName?: string,
      *   NextToken?: null|string,
      *   MaxResults?: null|int,
+     *   WorkGroup?: null|string,
      *   '@region'?: string|null,
      * } $input
      */
@@ -47,6 +56,7 @@ final class ListDatabasesInput extends Input
         $this->catalogName = $input['CatalogName'] ?? null;
         $this->nextToken = $input['NextToken'] ?? null;
         $this->maxResults = $input['MaxResults'] ?? null;
+        $this->workGroup = $input['WorkGroup'] ?? null;
         parent::__construct($input);
     }
 
@@ -55,6 +65,7 @@ final class ListDatabasesInput extends Input
      *   CatalogName?: string,
      *   NextToken?: null|string,
      *   MaxResults?: null|int,
+     *   WorkGroup?: null|string,
      *   '@region'?: string|null,
      * }|ListDatabasesInput $input
      */
@@ -76,6 +87,11 @@ final class ListDatabasesInput extends Input
     public function getNextToken(): ?string
     {
         return $this->nextToken;
+    }
+
+    public function getWorkGroup(): ?string
+    {
+        return $this->workGroup;
     }
 
     /**
@@ -124,6 +140,13 @@ final class ListDatabasesInput extends Input
         return $this;
     }
 
+    public function setWorkGroup(?string $value): self
+    {
+        $this->workGroup = $value;
+
+        return $this;
+    }
+
     private function requestBody(): array
     {
         $payload = [];
@@ -136,6 +159,9 @@ final class ListDatabasesInput extends Input
         }
         if (null !== $v = $this->maxResults) {
             $payload['MaxResults'] = $v;
+        }
+        if (null !== $v = $this->workGroup) {
+            $payload['WorkGroup'] = $v;
         }
 
         return $payload;